+++ inform/__init__.py
@@ -23,7 +23,7 @@ from .inform import (
     notify, debug, warn, error, fatal, panic,
 
     # the currently active informer
-    get_informer,
+    get_informer, set_informer,
 
     # culprit functions
     set_culprit, add_culprit, get_culprit, join_culprit